Added screenshot logic for failed tests

+++ src/test/java/stepDefinitions/Hooks.java
@@ -10,14 +10,13 @@ import com.aventstack.extentreports.reporter.ExtentSparkReporter;
 import com.aventstack.extentreports.reporter.configuration.Theme;
 import io.cucumber.java.*;
 import org.apache.commons.io.FileUtils;
+import org.openqa.selenium.OutputType;
+import org.openqa.selenium.TakesScreenshot;
 import org.openqa.selenium.WebDriver;
-import sun.reflect.misc.FieldUtil;
 import utility.TestContext;
 
 import java.io.File;
 import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 
@@ -26,6 +25,7 @@ public class Hooks {
     static ExtentReports extentReports;
     static ExtentTest extentTest;
     WebDriver driver;
+    static boolean screenshotExists = true;
 
     public Hooks(TestContext context) {
         driver = context.getWebDriverManager().getDriver();
@@ -49,16 +49,13 @@ public class Hooks {
 
         new File("reports").mkdir();
 
-        //create screenshots folder if doesn't exist
-        if (new File("screenshots").exists()) {
+        if (new File("screenshots").exists() && screenshotExists == true) {
             FileUtils.deleteDirectory(new File("screenshots"));
         }
-
-        new File("screenshots").mkdir();
     }
 
     @AfterAll
-    public static void tearDown() {
+    public static void tearDown() throws IOException {
         extentReports.flush();
     }
 
@@ -68,15 +65,29 @@ public class Hooks {
     }
 
     @After
-    public void afterTest(Scenario scenario) {
+    public void afterTest(Scenario scenario) throws IOException {
         if (scenario.getStatus() == Status.PASSED) {
             Markup m = MarkupHelper.createLabel("Test Passed", ExtentColor.GREEN);
             extentTest.log(com.aventstack.extentreports.Status.PASS, m);
         }
         else if (scenario.getStatus() == Status.FAILED) {
+            //create screenshots folder if doesn't exist
+            if (new File("screenshots").exists() && screenshotExists == true) {
+                FileUtils.deleteDirectory(new File("screenshots"));
+                screenshotExists = false;
+            }
+            new File("screenshots").mkdir();
+
+            //Capture screenshot for failed scenarios
+            TakesScreenshot ts = (TakesScreenshot)driver;
+            String screenshotPath = "screenshots/" + scenario.getName().replace(" ", "_") + ".png";
+            String screenshotPathToAttach = "../screenshots/" + scenario.getName().replace(" ", "_") + ".png";
+            File screenshotSource = ts.getScreenshotAs(OutputType.FILE);
+            FileUtils.copyFile(screenshotSource, new File(screenshotPath));
+
             Markup m = MarkupHelper.createLabel("Test Failed", ExtentColor.RED);
             extentTest.log(com.aventstack.extentreports.Status.FAIL, m);
-            extentTest.info("Screenshot attached", MediaEntityBuilder.createScreenCaptureFromPath("../screenshots/image.png").build());
+            extentTest.info("Screenshot attached", MediaEntityBuilder.createScreenCaptureFromPath(screenshotPathToAttach).build());
         }
         else if(scenario.getStatus() == Status.SKIPPED) {
             Markup m = MarkupHelper.createLabel("Test Skipped", ExtentColor.AMBER);
